+++ src/error.rs
@@ -2,13 +2,14 @@ use actix_web::body::BoxBody;
 use actix_web::http::StatusCode;
 use actix_web::{HttpResponse, ResponseError};
 use thiserror::Error;
+use crate::DeserializeError;
 
 #[derive(Debug, Error)]
 pub enum PayloadError {
     #[error("Payload error: {0}")]
     ActixPayload(#[from] actix_web::error::PayloadError),
     #[error("Error: {0}")]
-    Actix(#[from] actix_web::Error),
+    Deserialize(#[from] DeserializeError),
     #[error("Invalid content type")]
     InvalidContentType,
 }