+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.4.0
+- Removed actix-protobuf dependency
+- Dropped support for Paperclip
+- Add `serialize` and `deserialize` methods to public API.
+- Add XML support via the `xml` feature
+
 ## 0.3.0
 - Updated prost to 0.11
 - Updated actix-protobuf to 0.9